+++ src/utils.js
@@ -23,13 +23,6 @@ exports.findByName = (array, name, formerNames) => R.find((el) => {
   return false
 }, array)
 
-exports.filterByProp = R.curry((prop, props, array) => (
-  R.filter(R.pipe(
-    R.prop(prop),
-    R.includes(R.__, props),
-  ), array)
-))
-
 exports.loadJSON = (path, placeholders) => {
   let file = fs.readFileSync(path, 'utf-8')
   if (placeholders) {